requirements.txt
Django==1.8.4
Pillow==2.8.1
PyJWT==1.3.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
django-allauth==0.20.0
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.5
django-extra-fields==0.3
django-filter==0.10.0
django-guardian==1.3
django-haystack==2.4.0
django-libsass==0.3
django-model-utils==2.2
django-rest-auth==0.4.0
django-resto==1.2
djangorestframework==3.1.3
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.5.0
djoser==0.3.0
drf-extensions==0.2.7
drf-haystack==1.5.2
easy-thumbnails==2.2
jsonfield==1.0.3
libsass==0.8.2
oauthlib==0.7.2
psycopg2==2.6
pykurento==0.0.1
pysolr==3.3.2
python-dateutil==2.4.2
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.2.10
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
uWSGI==2.0.10
websocket-client==0.32.0
wheel==0.24.0
wsgiref==0.1.2